+++ life/src/main/java/org/yt/jr/projects/creatures/animals/carnivores/Wolf.java
@@ -1,4 +1,4 @@
-package org.yt.jr.projects.creatures.animals;
+package org.yt.jr.projects.creatures.animals.carnivores;
 
 import org.yt.jr.projects.creatures.Animal;
 import org.yt.jr.projects.creatures.CreatureType;
