+++ life/src/main/java/org/yt/jr/projects/creatures/animals/herbivores/Deer.java
@@ -1,4 +1,4 @@
-package org.yt.jr.projects.creatures.animals;
+package org.yt.jr.projects.creatures.animals.herbivores;
 
 import org.yt.jr.projects.creatures.Animal;
 import org.yt.jr.projects.creatures.CreatureType;
